+++ packages/css-color-parser/src/index.ts
@@ -1,24 +1,28 @@
-import { ColorData } from './color-data';
-import { ComponentValue, isFunctionNode, isTokenNode } from '@csstools/css-parser-algorithms';
+import type { ColorData } from './color-data';
+import type { ComponentValue } from '@csstools/css-parser-algorithms';
 import { TokenType } from '@csstools/css-tokenizer';
+import { color as colorFn } from './functions/color';
 import { colorKeyword } from './functions/color-keyword';
+import { colorMix } from './functions/color-mix';
 import { hex } from './functions/hex';
 import { hsl } from './functions/hsl';
 import { hwb } from './functions/hwb';
+import { isFunctionNode, isTokenNode } from '@csstools/css-parser-algorithms';
+import { lab } from './functions/lab';
+import { lch } from './functions/lch';
 import { namedColor } from './functions/named-color';
+import { oklab } from './functions/oklab';
+import { oklch } from './functions/oklch';
 import { rgb } from './functions/rgb';
 import { toLowerCaseAZ } from './util/to-lower-case-a-z';
-import { lab } from './functions/lab';
-import { colorMix } from './functions/color-mix';
-import { oklab } from './functions/oklab';
-import { lch } from './functions/lch';
-import { oklch } from './functions/oklch';
-import { color as colorFn } from './functions/color';
 
+export type { ColorData } from './color-data';
 export { ColorNotation } from './color-notation';
 export { SyntaxFlag } from './color-data';
-export { serializeRGB } from './serialize/rgb';
+export { colorDataTo } from './color-data';
 export { serializeP3 } from './serialize/p3';
+export { serializeRGB } from './serialize/rgb';
+
 
 export function color(colorNode: ComponentValue): ColorData | false {
 	if (isFunctionNode(colorNode)) {